+++ test/broker/c/plugin_evt_message_out.c
@@ -23,7 +23,7 @@ int callback_message_out(int event, void *event_data, void *user_data)
 	}
 	ed->topic = mosquitto_strdup("new-topic");
 	ed->payload = mosquitto_strdup("new-message");
-	ed->payloadlen = strlen(ed->payload);
+	ed->payloadlen = (uint32_t)strlen(ed->payload);
 	ed->properties = NULL;
 	if(mosquitto_property_add_string_pair(&ed->properties, MQTT_PROP_USER_PROPERTY, "key", "value")) abort();
